Gradient in env property

+++ models/macro_economics/substeps/utils.py
@@ -37,4 +37,4 @@ def initialize_id(shape, params):
     """
     Initialize a unique ID for each agent
     """
-    return torch.arange(0, shape[0]).reshape(-1, 1).int()
+    return torch.arange(0, shape[0]).reshape(-1, 1).float()
